# Float64.Add eligible for inlining with goto instead of for loop