@prashantv 2019-07-30
Thanks for your contribution @temoto 

I'd like to avoid complicating the code without some clear benefits, so ideally we'd want to see:
 - benchmark comparisons to see if there's an impactful difference
 - verification that this leads to inlined code

I'm not sure this will work since the Go compiler optimizations page calls out that inlining is disabled when complex features are used, and it includes loops as well as labels (which we're using for the goto here).

https://github.com/golang/go/wiki/CompilerOptimizations#function-inlining

> Only short and simple functions are inlined. To be inlined a function must contain less than ~40 expressions and does not contain complex things like function calls, loops, labels, closures, panic's, recover's, select's, switch'es, etc.