@temoto 2019-07-31
Yeah it may seem like unsupported dust. Copied this goto trick from pretty much same atomic float wrapper in my project, where I checked compiler output and was surprised to learn that `for` is a performance obstacle.

```
# before patch
./atomic.go:288:6: cannot inline (*Float64).Add: unhandled op FOR

# after patch
./atomic.go:288:6: can inline (*Float64).Add as: method(*Float64) func(float64) float64 { tryCAS: ; old := f.Load(); new := old + s; if f.CAS(old, new) { return new }; goto tryCAS }
./atomic.go:300:14: inlining call to (*Float64).Add method(*Float64) func(float64) float64 { tryCAS: ; old := f.Load(); new := old + s; if f.CAS(old, new) { return new }; goto tryCAS }
```

```
# benchcmp
benchmark                              old ns/op     new ns/op     delta
BenchmarkStress/f64/parallel-4         102           98.9          -3.04%
BenchmarkStress/f64/serial-4           45.1          44.7          -0.93%
BenchmarkStress/f64/add/serial-4       13.8          13.4          -2.61%
BenchmarkStress/f64/add/parallel-4     35.6          37.5          +5.11%
```

The f64/add code, I doubt it's useful to include:
```go
func stressFloat64Add() func() {
       var atom Float64
       return func() {
               atom.Add(1.1)
       }
}
```