@prashantv 2019-08-22
Thanks for the additional information @temoto

I was able to verify that `goto` causes the function to be inlined. I created a branch with both versions, where `AddLoop` uses the old way, while `Add` is using the goto.

We see that `AddLoop` has a function call:
```
     280ms      280ms     68a605: MOVSD_XMM $f64.3ff199999999999a(SB), X0 ;go.uber.org/atomic.stressFloat64AddLoop.func1 stress_test.go:65
         .          .     68a60d: MOVSD_XMM X0, 0x8(SP)                   ;stress_test.go:65
                                                                          ;go.uber.org/atomic.stressFloat64AddLoop.func1 stress_test.go:65
      10ms     26.43s     68a613: CALL go.uber.org/atomic.(*Float64).AddLoop(SB)
         .          .     68a618: MOVQ 0x18(SP), BP                       ;stress_test.go:66
```

On the other hand, `Add` does not:
```
     790ms      790ms     68a5ae: MOVSD_XMM $f64.3ff199999999999a(SB), X1 ;go.uber.org/atomic.(*Float64).Add atomic.go:291
         .          .     68a5b6: ADDSD X1, X0                            ;atomic.go:291
     1.35s     18.83s     68a5ba: NOPL                                    ;go.uber.org/atomic.(*Float64).Add atomic.go:292
         .      7.08s     68a5bb: NOPL                                    ;go.uber.org/atomic.(*Float64).Load atomic.go:279
         .          .     68a5bc: NOPL                                    ;atomic.go:316
         .      360ms     68a5bd: NOPL                                    ;go.uber.org/atomic.(*Float64).CAS atomic.go:316
     360ms      360ms     68a5be: MOVQ X0, BX                             ;math.Float64bits unsafe.go:23
     330ms      330ms     68a5c3: MOVQ DX, AX                             ;go.uber.org/atomic.(*Float64).CAS atomic.go:316
         .          .     68a5c6: LOCK CMPXCHGQ BX, 0(CX)                 ;atomic.go:316
    16.74s     16.74s     68a5cb: SETE DL                                 ;go.uber.org/atomic.(*Float64).CAS atomic.go:316
      50ms       50ms     68a5ce: TESTL DL, DL
         .          .     68a5d0: JE 0x68a5a5                             ;atomic.go:292

```

However, the results of the benchmark didn't really show any improvement,
```
BenchmarkStress/f64-add-loop/serial-4   1000000000               9.17 ns/op
BenchmarkStress/f64-add-loop/parallel-4                 100000000               73.2 ns/op
BenchmarkStress/f64-add/serial-4                        1000000000               9.36 ns/op
BenchmarkStress/f64-add/parallel-4                      100000000               72.3 ns/op
PASS
ok      go.uber.org/atomic      35.235s
```

I ran benchmarks with longer `benchtime` and saw the same -- the `AddLoop` was actually faster (in serial),
```
BenchmarkStress/f64-add/serial-4        3000000000               9.65 ns/op
BenchmarkStress/f64-add/parallel-4              500000000               76.0 ns/op
BenchmarkStress/f64-add-loop/serial-4           3000000000               9.53 ns/op
BenchmarkStress/f64-add-loop/parallel-4         500000000               77.0 ns/op
PASS
ok      go.uber.org/atomic      151.635s
```

I'm a little surprised that the inlining seems to have a slight negative impact when run serially.  The parallel benchmark varied -- sometimes the loop version was faster, sometimes the goto version was faster.

cc @abhinav for his thoughts